requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
chromadb==0.4.15
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==0.8.11
markdown==3.4.4
streamlit==1.29.0
transformers==4.30.0
torch==2.0.0